Moved CalcSettings to own file
